requirements.txt
aws-cdk-lib==2.9.0
black==21.12b0
boto3==1.20.44
constructs>=10.0.0,<11.0.0
pandas==1.3.5
pylint==2.12.2
pytest==6.2.5
requests==2.27.1
